Refactor to look for instructions in subfolder
* Save the base filename in options
* Look in subfolder for load/save
* Moved helpers to instructions.py to remove GUI dependency
* Modified makedistro scripts to package instructions folder
