+++ scripts/makedistro.bat
@@ -6,4 +6,4 @@ pip install --upgrade -r requirements.txt
 pip install --upgrade openai
 pip install --upgrade google.generativeai
 pip install --upgrade anthropic
-pyinstaller --noconfirm --additional-hooks-dir="PySubtitleHooks" --add-data "theme/*;theme/" --add-data "instructions*;." --add-data "LICENSE;." --add-data "assets/gui-subtrans.ico;." gui-subtrans.py
+pyinstaller --noconfirm --additional-hooks-dir="PySubtitleHooks" --add-data "theme/*;theme/" --add-data "instructions*;instructions/" --add-data "LICENSE;." --add-data "assets/gui-subtrans.ico;." gui-subtrans.py
